+++ home/tests/test_team_formation.py
@@ -8,8 +8,23 @@ from django.urls import reverse
 from accounts.factories import UserAvailabilityFactory, UserFactory
 from accounts.models import CustomUser, UserAvailability
 from home.admin import SessionAdmin
-from home.factories import ProjectFactory, TeamFactory
-from home.forms import ApplicantFilterForm
+from home.availability import (
+    calculate_overlap,
+    calculate_team_overlap,
+    count_one_hour_blocks,
+    format_slot_as_time,
+    format_slots_as_ranges,
+)
+from home.factories import (
+    TeamFactory,
+    ProjectFactory,
+    SurveyFactory,
+    SessionMembershipFactory,
+    ProjectPreferenceFactory,
+    UserSurveyResponseFactory,
+)
+from home.filters import ApplicantFilterSet
+from home.forms import ApplicantFilterForm, OverlapAnalysisForm
 from home.models import (
     ProjectPreference,
     Question,
@@ -20,13 +35,8 @@ from home.models import (
     TypeField,
     UserSurveyResponse,
 )
-from home.availability import (
-    calculate_overlap,
-    calculate_team_overlap,
-    count_one_hour_blocks,
-    format_slot_as_time,
-    format_slots_as_ranges,
-)
+
+from home.views.team_formation import get_teams_with_statistics
 
 
 class AvailabilityUtilsTestCase(TestCase):
@@ -166,6 +176,7 @@ class ApplicantFilterFormTestCase(TestCase):
             application_start_date="2025-01-15",
             application_end_date="2025-02-15",
         )
+        self.survey = SurveyFactory(session=self.session)
 
         self.team = TeamFactory(session=self.session, name="Test Team")
 
@@ -186,6 +197,41 @@ class ApplicantFilterFormTestCase(TestCase):
         }
         form = ApplicantFilterForm(data=form_data, session=self.session)
         self.assertTrue(form.is_valid())
+
+    def test_project_preference_filter(self):
+        """Test filtering applicants by project preference."""
+        project1 = ProjectFactory(name="Project 1")
+        project2 = ProjectFactory(name="Project 2")
+        self.session.available_projects.add(project1, project2)
+
+        applicant1 = UserFactory(username="app1")
+        UserSurveyResponseFactory(user=applicant1, survey=self.survey)
+        ProjectPreferenceFactory(
+            user=applicant1, session=self.session, project=project1
+        )
+
+        applicant2 = UserFactory(username="app2")
+        UserSurveyResponseFactory(user=applicant2, survey=self.survey)
+        ProjectPreferenceFactory(
+            user=applicant2, session=self.session, project=project2
+        )
+
+        qs = UserSurveyResponse.objects.filter(survey=self.survey)
+
+        # Filter by Project 1
+        filterset = ApplicantFilterSet(
+            data={"project_preferences": project1.id}, queryset=qs, session=self.session
+        )
+        self.assertTrue(filterset.is_valid())
+        self.assertEqual(filterset.qs.count(), 1)
+        self.assertEqual(filterset.qs.first().user, applicant1)
+
+        # Filter by Project 2
+        filterset = ApplicantFilterSet(
+            data={"project_preferences": project2.id}, queryset=qs, session=self.session
+        )
+        self.assertEqual(filterset.qs.count(), 1)
+        self.assertEqual(filterset.qs.first().user, applicant2)
 
 
 class TeamFormationViewTestCase(TestCase):
@@ -612,3 +658,102 @@ class TeamFormationViewTestCase(TestCase):
         self.assertIn("Project Alpha", content)
         self.assertIn("Project Beta", content)
         self.assertIn("Project Gamma", content)
+
+    def test_overlap_includes_existing_djangonauts(self):
+        """Test that overlap calculation includes existing djangonauts."""
+        # Create a djangonaut on the team
+        djangonaut = UserFactory(username="djangonaut")
+        team = TeamFactory(session=self.session, name="Test Team")
+        SessionMembershipFactory(
+            user=djangonaut,
+            session=self.session,
+            team=team,
+            role=SessionMembership.DJANGONAUT,
+        )
+        # Give djangonaut availability (slot 24.0 = Mon 00:00)
+        UserAvailabilityFactory(user=djangonaut, slots=[24.0, 24.5])
+
+        # Create a navigator on the team
+        navigator = UserFactory(username="navigator")
+        SessionMembershipFactory(
+            user=navigator,
+            session=self.session,
+            team=team,
+            role=SessionMembership.NAVIGATOR,
+        )
+        # Give navigator availability (slot 24.0 = Mon 00:00)
+        UserAvailabilityFactory(user=navigator, slots=[24.0, 24.5])
+
+        # Create an applicant with matching availability
+        applicant = UserFactory(username="applicant")
+        UserAvailabilityFactory(user=applicant, slots=[24.0, 24.5])
+
+        # We need "selected_users" for the form. The form uses cleaned_data['user_ids']
+        # But for testing `calculate_navigator_overlap_context` we can mock or use valid data
+
+        form_data = {
+            "overlap-team": team.id,
+            "overlap-analysis_type": "overlap-navigator",
+            "overlap-user_ids": str(applicant.id),
+        }
+        form = OverlapAnalysisForm(data=form_data, session=self.session)
+        self.assertTrue(form.is_valid())
+
+        context = form.calculate_navigator_overlap_context()
+
+        # Check that existing djangonauts are in context
+        self.assertIn("existing_djangonauts", context)
+        self.assertEqual(len(context["existing_djangonauts"]), 1)
+        self.assertEqual(context["existing_djangonauts"][0], djangonaut)
+
+        # Check overlap calculation (should be 1 hour because all 3 overlap)
+        # The overlap logic calculates intersection of ALL users.
+        # Navigator and Djangonaut and Applicant all have [24.0, 24.5].
+        self.assertEqual(context["hour_blocks"], 1)
+
+    def test_compare_availability_url_in_context(self):
+        """Test that compare_availability_url is present in overlap context."""
+        navigator = UserFactory(username="navigator")
+        team = TeamFactory(session=self.session, name="Test Team")
+        SessionMembershipFactory(
+            user=navigator,
+            session=self.session,
+            team=team,
+            role=SessionMembership.NAVIGATOR,
+        )
+
+        applicant = UserFactory(username="applicant")
+
+        form_data = {
+            "overlap-team": team.id,
+            "overlap-analysis_type": "overlap-navigator",
+            "overlap-user_ids": str(applicant.id),
+        }
+        form = OverlapAnalysisForm(data=form_data, session=self.session)
+        self.assertTrue(form.is_valid())
+
+        context = form.calculate_navigator_overlap_context()
+        self.assertIn("compare_availability_url", context)
+        url = context["compare_availability_url"]
+        self.assertIn(reverse("compare_availability"), url)
+        self.assertIn(str(navigator.id), url)
+        self.assertIn(str(applicant.id), url)
+
+    def test_team_statistics_has_compare_url(self):
+        """Test that team statistics includes compare availability URL."""
+        djangonaut = UserFactory(username="djangonaut")
+        team = TeamFactory(session=self.session, name="Test Team")
+        SessionMembershipFactory(
+            user=djangonaut,
+            session=self.session,
+            team=team,
+            role=SessionMembership.DJANGONAUT,
+        )
+
+        teams_data = get_teams_with_statistics(self.session)
+        self.assertEqual(len(teams_data), 1)
+        stat = teams_data[0]
+
+        self.assertTrue(hasattr(stat, "compare_availability_url"))
+        self.assertIn(reverse("compare_availability"), stat.compare_availability_url)
+        self.assertIn(str(djangonaut.id), stat.compare_availability_url)
